Initial commit
A perl program to create a maze on the command line. It then drops a mouse and cheese in the maze. The mouse then solves the maze by finding the cheese. It requires two numbers to set the size of the maze and then the affinity the mouse has for finding the maze expressed as a decimal < 1.